fix: skip owner/group lookup on Windows — user.Lookup returns SIDs
On Windows, user.Lookup and user.LookupGroup return SID strings
(e.g. S-1-5-21-...) as the Uid/Gid fields. strconv.Atoi fails on
these, causing an error before any template can be processed when
owner or group is set in the resource config.

Since chownFile is already a no-op on Windows, the uid/gid values
are never applied. Move resolveOwnership to platform-specific files:
resource_unix.go performs the lookup as before; resource_windows.go
returns defaults and logs a warning if owner/group are configured.

Add TestResolveOwnership_Defaults and TestResolveOwnership_ExplicitValues
to cover the platform-specific paths on all CI runners.

Update CHANGELOG for #574.
