+++ pkg/template/resource.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"os/user"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -109,43 +108,6 @@ type TemplateResource struct {
 
 var ErrEmptySrc = errors.New("empty src template")
 
-// resolveOwnership resolves the UID and GID for the template resource.
-// If Owner/Group are specified, it looks them up. Otherwise uses effective UID/GID.
-func resolveOwnership(owner, group string, uid, gid int) (int, int, error) {
-	// Resolve UID
-	if uid == -1 {
-		if owner != "" {
-			u, err := user.Lookup(owner)
-			if err != nil {
-				return 0, 0, fmt.Errorf("cannot find owner's UID: %w", err)
-			}
-			uid, err = strconv.Atoi(u.Uid)
-			if err != nil {
-				return 0, 0, fmt.Errorf("cannot convert string to int: %w", err)
-			}
-		} else {
-			uid = os.Geteuid()
-		}
-	}
-
-	// Resolve GID
-	if gid == -1 {
-		if group != "" {
-			g, err := user.LookupGroup(group)
-			if err != nil {
-				return 0, 0, fmt.Errorf("Cannot find group's GID - %s", err.Error())
-			}
-			gid, err = strconv.Atoi(g.Gid)
-			if err != nil {
-				return 0, 0, fmt.Errorf("Cannot convert string to int - %s", err.Error())
-			}
-		} else {
-			gid = os.Getegid()
-		}
-	}
-
-	return uid, gid, nil
-}
 
 // normalizePrefix concatenates global config prefix with resource prefix.
 // This allows hierarchical prefixes like /production/myapp where
